requirements.txt
stockfish==3.14.0
selenium==3.8.1
chess==1.5.0
PyQt5
pywin32==307
chromedriver_autoinstaller